+++ client/src/auth/ProtectedRoutes/ProtectedRoutes.tsx
@@ -2,11 +2,18 @@ import React, { useEffect } from "react";
 import { Navigate, Outlet } from "react-router-dom";
 import { useAuth } from "@hooks";
 
+/**
+ * Validates session cookie when visiting one of these routes.
+ *
+ * This differs fromm AutoLoginRoutes because we redirect *away*
+ * from protected routes, vs *to* them.
+ */
 export default function PrivateRoute(): React.JSX.Element {
   const { session, validateSession } = useAuth();
 
   useEffect(() => {
-    // If `document.cookie` exists, we need to validate before letting them through.
+    // If `document.cookie` exists but no session in state,
+    // we need to validate before letting them through.
     if (document.cookie && !session) {
       (async (): Promise<void> => {
         await validateSession();
@@ -14,14 +21,18 @@ export default function PrivateRoute(): React.JSX.Element {
     }
   }, [validateSession, session]);
 
+  // If user has a cookie but no session stored
+  // in our state, we need to validate the cookie.
   if (document.cookie && !session) {
-    console.log({ from: "ProtectedRoutes Component", status: "document.cookie && !session", returning: "<div>Loading...</div>", session });
     return <div>Loading...</div>;
   }
+  // If there is no cookie and no session, it
+  // means the user has to authenticate manually.
   if (!document.cookie && !session) {
-    console.log({ from: "ProtectedRoutes Component", status: "!document.cookie && !session", returning: "<Navigate to='/' />", session });
     return <Navigate to="/" />;
   }
-  console.log({ from: "ProtectedRoutes Component", returning: "<Outlet />", session });
+  // If there is a cookie (would have been verified
+  // by this point) and there is a session let the
+  // "request" continue along the route.
   return <Outlet />;
 }
