+++ client/src/hooks/useFirstChildShouldBe.tsx
@@ -1,7 +1,7 @@
-import React, { ReactNode } from "react";
+import React, { ReactNode, isValidElement } from "react";
 
 // Custom hook to validate if the first child is a specific component
 export default function useFirstChildShouldBe<T>(children: ReactNode, shouldBeComponent: T): boolean {
   const firstChild = React.Children.toArray(children)[0];
-  return React.isValidElement(firstChild) && firstChild.type === shouldBeComponent;
+  return isValidElement(firstChild) && firstChild.type === shouldBeComponent;
 }
